Consolidate cluster tests
