+++ cluster_test.go
@@ -1,0 +1,211 @@
+package jobsd
+
+import (
+	"strconv"
+	"sync"
+	"sync/atomic"
+	"testing"
+	"time"
+
+	"github.com/simpleframeworks/testc"
+	"github.com/sirupsen/logrus"
+)
+
+func TestRunOnceCreatorUnique(test *testing.T) {
+	t := testc.New(test)
+
+	t.Given("a JobsD instance")
+	jd := testSetup(logrus.ErrorLevel)
+
+	nodes := 10
+	t.Given("a " + strconv.Itoa(nodes) + " JobsD instance cluster with 5 workers each")
+	jdInstances := []*JobsD{jd}
+	for i := 0; i < nodes-1; i++ {
+		jdInstances = append(jdInstances, New(jd.GetDB()).Logger(jd.GetLogger()).WorkerNum(5))
+	}
+
+	runTime := 800 * time.Millisecond
+	t.Given("a job that increments a counter and takes " + runTime.String())
+	wait := sync.WaitGroup{}
+	var runCounter uint32
+	jobFunc := func() error {
+		defer wait.Done()
+		atomic.AddUint32(&runCounter, 1)
+		<-time.After(runTime)
+		return nil
+	}
+
+	t.Given("the instances can run the job")
+	for _, qInst := range jdInstances {
+		qInst.RegisterJob("jobName", jobFunc)
+	}
+
+	t.When("we bring up the JobsD instances")
+	for _, qInst := range jdInstances {
+		t.Assert.NoError(qInst.Up())
+	}
+
+	t.When("we run the same unique job run on each of the instances in the cluster")
+	wait.Add(1) //we only expect it to run once
+	for _, qInst := range jdInstances {
+		_, err := qInst.CreateRun("jobName").Unique("UniqueJobName").Run()
+		t.Assert.NoError(err)
+	}
+
+	t.When("we wait until it finishes")
+	wait.Wait()
+
+	t.Then("the job should have run only once")
+	t.Assert.Equal(1, int(runCounter))
+
+	waitTime := 500 * time.Millisecond
+	t.When("we wait " + waitTime.String())
+	<-time.After(waitTime)
+
+	t.Then("the job should have still only run once")
+	t.Assert.Equal(1, int(runCounter))
+
+	for _, qInst := range jdInstances {
+		t.Assert.NoError(qInst.Down())
+	}
+	testTeardown(jd)
+}
+
+func TestRunScheduleCreatorUnique(test *testing.T) {
+	t := testc.New(test)
+
+	jobName := "TestRunScheduleCreatorUnique"
+
+	t.Given("a JobsD instance")
+	jd := testSetup(logrus.ErrorLevel)
+
+	nodes := 5
+	t.Given("a " + strconv.Itoa(nodes) + " JobsD instance cluster with 2 workers each")
+	jdInstances := []*JobsD{jd}
+	for i := 0; i < nodes-1; i++ {
+		jdInstances = append(jdInstances, New(jd.GetDB()).Logger(jd.GetLogger()).WorkerNum(2).RunTimeout(time.Second*30))
+	}
+
+	runTime := 500 * time.Millisecond
+	t.Given("a job that increments a counter and takes " + runTime.String())
+	wait := sync.WaitGroup{}
+	var runCounter uint32
+	jobFunc := func() error {
+		defer wait.Done()
+		atomic.AddUint32(&runCounter, 1)
+		<-time.After(runTime)
+		return nil
+	}
+
+	t.Given("the instances can run the job")
+	for _, qInst := range jdInstances {
+		qInst.RegisterJob(jobName, jobFunc)
+	}
+
+	interval := 200 * time.Millisecond
+	t.Given("a schedule that runs at set uniform interval of " + interval.String())
+	triggerTime := time.Now()
+	scheduleFunc := func(now time.Time) time.Time {
+		return triggerTime.Add(interval)
+	}
+
+	t.Given("the instances can use the schedule")
+	for _, qInst := range jdInstances {
+		qInst.RegisterSchedule("scheduleName", scheduleFunc)
+	}
+
+	t.When("we bring up the JobsD instances")
+	for _, qInst := range jdInstances {
+		t.Assert.NoError(qInst.Up())
+	}
+
+	t.When("we run the same unique job run on each of the instances in the cluster")
+	wait.Add(2) //we only expect it to run twice
+	for _, qInst := range jdInstances {
+		_, err := qInst.CreateRun(jobName).Schedule("scheduleName").Unique(jobName + "UniqueJobName").Limit(2).Run()
+		t.Assert.NoError(err)
+	}
+
+	t.When("we wait until it finishes")
+	wait.Wait()
+
+	t.Then("the job should have run twice")
+	t.Assert.Equal(2, int(runCounter))
+
+	waitTime := 500 * time.Millisecond
+	t.When("we wait " + waitTime.String())
+	<-time.After(waitTime)
+
+	t.Then("the job should have still only run twice")
+	t.Assert.Equal(2, int(runCounter))
+
+	for _, qInst := range jdInstances {
+		t.Assert.NoError(qInst.Down())
+	}
+}
+
+func TestJobsDClusterWorkSharing(test *testing.T) {
+	t := testc.New(test)
+
+	jobName := "TestJobsDClusterWorkSharing" // Must be unique otherwise tests may collide
+
+	t.Given("a JobsD instance")
+	jd := testSetup(logrus.ErrorLevel)
+
+	nodes := 10
+	t.Given("a " + strconv.Itoa(nodes) + " JobsD instance cluster with one worker each")
+	jdInstances := []*JobsD{jd}
+	for i := 0; i < nodes-1; i++ {
+		jdInstances = append(jdInstances, New(jd.GetDB()).Logger(jd.GetLogger()).WorkerNum(1).PollInterval(100*time.Millisecond))
+	}
+
+	runTime := 150 * time.Millisecond
+	t.Given("a job that increments a counter and takes " + runTime.String())
+	wait := sync.WaitGroup{}
+	var runCounter uint32
+	jobFunc := func() error {
+		defer wait.Done()
+		atomic.AddUint32(&runCounter, 1)
+		<-time.After(runTime)
+		return nil
+	}
+
+	t.Given("the cluster can run the job")
+	for _, qInst := range jdInstances {
+		qInst.RegisterJob(jobName, jobFunc)
+	}
+
+	t.When("we bring up the JobsD instances")
+	for _, qInst := range jdInstances {
+		t.Assert.NoError(qInst.Up())
+	}
+
+	runs := nodes * 3
+	runIDs := []int64{}
+	t.When("we run a job " + strconv.Itoa(runs) + " times from the first instance in the cluster")
+	for i := 0; i < runs; i++ {
+		wait.Add(1)
+		runID, err := jdInstances[0].CreateRun(jobName).Run()
+		t.Assert.NoError(err)
+		runIDs = append(runIDs, runID)
+	}
+
+	t.Then("the job should have run " + strconv.Itoa(runs) + " times")
+	wait.Wait()
+	t.Assert.Equal(runs, int(runCounter))
+
+	t.Then("the job runs should have been distributed across the cluster and run")
+	nonLocalRuns := 0
+	for _, runID := range runIDs {
+		theState := jdInstances[0].GetRunState(runID)
+		if theState.RunStartedBy != nil && *theState.RunStartedBy != jdInstances[0].instance.ID {
+			nonLocalRuns++
+		}
+	}
+	t.Assert.Greater(nonLocalRuns, 1)
+
+	for _, qInst := range jdInstances {
+		t.Assert.NoError(qInst.Down())
+	}
+	testTeardown(jd)
+}
