+++ jobsD_test.go
@@ -3,7 +3,6 @@ package jobsd
 import (
 	"errors"
 	"fmt"
-	"strconv"
 	"sync"
 	"sync/atomic"
 	"testing"
@@ -344,71 +343,6 @@ func TestJobsDScheduledRunMulti(test *testing.T) {
 	testTeardown(jd)
 }
 
-func TestJobsDClusterWorkSharing(test *testing.T) {
-	t := testc.New(test)
-
-	jobName := "TestJobsDClusterWorkSharing" // Must be unique otherwise tests may collide
-
-	t.Given("a JobsD instance")
-	jd := testSetup(logrus.ErrorLevel)
-
-	nodes := 10
-	t.Given("a " + strconv.Itoa(nodes) + " JobsD instance cluster with one worker each")
-	jdInstances := []*JobsD{jd}
-	for i := 0; i < nodes-1; i++ {
-		jdInstances = append(jdInstances, New(jd.GetDB()).Logger(jd.GetLogger()).WorkerNum(1).PollInterval(100*time.Millisecond))
-	}
-
-	runTime := 150 * time.Millisecond
-	t.Given("a job that increments a counter and takes " + runTime.String())
-	wait := sync.WaitGroup{}
-	var runCounter uint32
-	jobFunc := func() error {
-		defer wait.Done()
-		atomic.AddUint32(&runCounter, 1)
-		<-time.After(runTime)
-		return nil
-	}
-
-	t.Given("the cluster can run the job")
-	for _, qInst := range jdInstances {
-		qInst.RegisterJob(jobName, jobFunc)
-	}
-
-	t.When("we bring up the JobsD instances")
-	for _, qInst := range jdInstances {
-		t.Assert.NoError(qInst.Up())
-	}
-
-	runs := nodes * 3
-	runIDs := []int64{}
-	t.When("we run a job " + strconv.Itoa(runs) + " times from the first instance in the cluster")
-	for i := 0; i < runs; i++ {
-		wait.Add(1)
-		runID, err := jdInstances[0].CreateRun(jobName).Run()
-		t.Assert.NoError(err)
-		runIDs = append(runIDs, runID)
-	}
-
-	t.Then("the job should have run " + strconv.Itoa(runs) + " times")
-	wait.Wait()
-	t.Assert.Equal(runs, int(runCounter))
-
-	t.Then("the job runs should have been distributed across the cluster and run")
-	nonLocalRuns := 0
-	for _, runID := range runIDs {
-		theState := jdInstances[0].GetRunState(runID)
-		if theState.RunStartedBy != nil && *theState.RunStartedBy != jdInstances[0].instance.ID {
-			nonLocalRuns++
-		}
-	}
-	t.Assert.Greater(nonLocalRuns, 1)
-
-	for _, qInst := range jdInstances {
-		t.Assert.NoError(qInst.Down())
-	}
-	testTeardown(jd)
-}
 func ExampleJobsD() {
 	jobName := "ExampleJobsD" // Must be unique otherwise tests may collide
